Simplify the repository and fix xtask config
This change simplifies the repository and fixes the xtask config. The
config now allows you to run xtask from any directory within the
repository. In addition, the xtask crate no longer has third party
dependencies and all logic is in one file. The diagram has been
updated accordingly.

The reasoning to use xtask is now more concise, ".gitignore" only
contains what is minimally needed, and the config file itself is
has been renamed to match present day Rust standards. The edition is
now "2024" with the workspace resolver moving to "3" alongside it.

This changes are all in service of simplifying and modernizing the
repository.

Signed-off-by: Nick Gerace <nickagerace@gmail.com>