Simplify further by moving all crates to root
This change simplifies things further by bringing all crates to the
root of the repository. While this could get messy with real world
use, it makes things easier to read for the repository without a layer
of redirection. This change also fixes the "config.toml" rename in the
diagram.

Included are changes to the build scripts and a fix to the diagram
visually.

Signed-off-by: Nick Gerace <nickagerace@gmail.com>